Modified Add shipment tracking activity to fragment - Part I
