+++ WooCommerce/src/main/AndroidManifest.xml
@@ -74,9 +74,6 @@
             android:name=".ui.imageviewer.ImageViewerActivity"
             android:theme="@style/ImageViewer.Activity"/>
 
-        <activity
-            android:name=".ui.orders.AddOrderShipmentTrackingActivity"/>
-
         <!-- Services -->
         <service
             android:name=".push.FCMRegistrationIntentService"
