+++ WooCommerce/src/main/kotlin/com/woocommerce/android/di/ActivityBindingModule.kt
@@ -10,7 +10,6 @@ import com.woocommerce.android.ui.main.MainModule
 import com.woocommerce.android.ui.notifications.NotifsListModule
 import com.woocommerce.android.ui.notifications.ReviewDetailModule
 import com.woocommerce.android.ui.orders.AddOrderNoteModule
-import com.woocommerce.android.ui.orders.AddOrderShipmentTrackingActivity
 import com.woocommerce.android.ui.orders.AddOrderShipmentTrackingModule
 import com.woocommerce.android.ui.orders.AddOrderTrackingProviderListModule
 import com.woocommerce.android.ui.orders.OrderDetailModule
@@ -42,7 +41,9 @@ abstract class ActivityBindingModule {
             AddOrderNoteModule::class,
             NotifsListModule::class,
             ReviewDetailModule::class,
-            SitePickerModule::class))
+            SitePickerModule::class,
+            AddOrderShipmentTrackingModule::class,
+            AddOrderTrackingProviderListModule::class))
     abstract fun provideMainActivityInjector(): MainActivity
 
     @ActivityScope
@@ -71,11 +72,4 @@ abstract class ActivityBindingModule {
     @ActivityScope
     @ContributesAndroidInjector(modules = arrayOf(ProductDetailModule::class))
     abstract fun provideProductDetailActivityInjector(): ProductDetailActivity
-
-    @ActivityScope
-    @ContributesAndroidInjector(modules = [
-        AddOrderShipmentTrackingModule::class,
-        AddOrderTrackingProviderListModule::class
-    ])
-    abstract fun provideAddOrderShipmentTrackingActivity(): AddOrderShipmentTrackingActivity
 }
