+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/base/TopLevelFragmentView.kt
@@ -7,11 +7,13 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.woocommerce.android.ui.orders.AddOrderNoteFragment
+import com.woocommerce.android.ui.orders.AddOrderShipmentTrackingFragment
 import com.woocommerce.android.ui.orders.OrderDetailFragment
 import com.woocommerce.android.ui.orders.OrderFulfillmentFragment
 import com.woocommerce.android.ui.orders.OrderProductListFragment
 import com.woocommerce.android.ui.orders.OrdersViewRouter
 import org.wordpress.android.fluxc.model.WCOrderModel
+import org.wordpress.android.fluxc.model.order.OrderIdentifier
 
 /**
  * Special interface for top-level fragments like those hosted by the bottom bar.
@@ -121,6 +123,21 @@ interface TopLevelFragmentView : FragmentManager.OnBackStackChangedListener, Ord
         val tag = AddOrderNoteFragment.TAG
         if (!popToState(tag)) {
             loadChildFragment(AddOrderNoteFragment.newInstance(order), tag)
+        }
+    }
+
+    override fun openAddOrderShipmentTracking(
+        orderIdentifier: OrderIdentifier,
+        orderTrackingProvider: String,
+        isCustomProvider: Boolean
+    ) {
+        val tag = AddOrderShipmentTrackingFragment.TAG
+        if (!popToState(tag)) {
+            loadChildFragment(AddOrderShipmentTrackingFragment.newInstance(
+                    orderIdentifier,
+                    orderTrackingProvider,
+                    isCustomProvider
+            ), tag)
         }
     }
 }
