+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/OrdersViewRouter.kt
@@ -1,6 +1,7 @@
 package com.woocommerce.android.ui.orders
 
 import org.wordpress.android.fluxc.model.WCOrderModel
+import org.wordpress.android.fluxc.model.order.OrderIdentifier
 
 interface OrdersViewRouter {
     fun openOrderDetail(order: WCOrderModel, markOrderComplete: Boolean = false)
@@ -8,4 +9,9 @@ interface OrdersViewRouter {
     fun openOrderFulfillment(order: WCOrderModel, isUsingCachedShipmentTrackings: Boolean = false)
     fun openOrderProductList(order: WCOrderModel)
     fun openAddOrderNote(order: WCOrderModel)
+    fun openAddOrderShipmentTracking(
+        orderIdentifier: OrderIdentifier,
+        orderTrackingProvider: String,
+        isCustomProvider: Boolean
+    )
 }
